added a guessing game

+++ src/main/java/org/example/Lab4/Circle.java
@@ -8,7 +8,7 @@ public class Circle extends Shape
     public Circle() {
     }
 
-    public Circle(int radius) {
+    public Circle(double radius) {
         this.radius = radius;
     }
